Create MotionGraphicsPlayground
A rather extensive bit of code I've been working on. I can create 10 types of shapes and alter their size and position and give them nine adjustable behaviors. This program started as a way to make black and white shape masks for further work, but I added an array of color as well. Next video will be black and white setup of these shapes. Made with #processing and some help from #chatgpt.